# WAIL OUAISSA GraphQL project 2/WAIL OUAISSA GraphQL Server/src/main/java/com/example/DemoGraphQL/model/Tag.java
package com.example.DemoGraphQL.model;

import lombok.Data;
import lombok.Getter;
import lombok.Setter;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;

@Entity
@Getter
@Setter
public class Tag {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private String id;

    private String name;
}